Refactor domains and add function to find subdomain
